// src/common/enums/competition-type.enum.ts
export enum CompetitionType {
  WEEKEND_LEAGUE = "weekend_league",
  CHAMPIONS_LEAGUE = "champions_league",
  EUROPA_LEAGUE = "europa_league",
  AFCON = "afcon",
  COPA_AMERICA = "copa_america",
  EUROS = "euros",
  UEFA_NATIONS_LEAGUE = "uefa_nations_league",
  INTERNATIONAL_FRIENDLY = "international_friendly",
  INTERNATIONAL_QUALIFIER = "international_qualifier",
  OTHER = "other",
}